Project inception: transform for JakartaEE 10 and for Java 11
The goal of the project is to transform Apache Jackrabbit JCR Server to
be compliant with JakartaEE 10.
For doing, the project embeds the source code of Apache Jackrabbit JCR
Server and uses the OpenWrite Maven plugin to transform the code for
JakartaEE and for Java 11. The code has required to be touched by hand
to fix the changes in the JakartaEE 10's Servlet API.